refactor(core): centralize checked model type validation
