+++ rust/fsl-core/src/public_kernel.rs
@@ -6,10 +6,18 @@ use std::collections::{BTreeMap, BTreeSet, HashMap};
 use std::fmt;
 
 use fsl_syntax::{
-    AggregateKind, Binder, ConditionalSpans, Expr, LValue, MetaTag, Pattern, SourcePos, Span,
-    SpecItem, Statement,
+    AggregateKind, Binder, ConditionalSpans, Expr, LValue, MetaTag, Pattern, Span, SpecItem,
+    Statement,
 };
 use serde_json::{Map, Value, json};
+
+use crate::typecheck::{
+    TypeEnv, TypecheckError, aggregate_condition, base_env, binary_operand_types, binder_type,
+    ensure_assignable, expression_type, extend_pattern_binding, finite_binder_candidates,
+    index_key_type, infer_type, lvalue_type, method_argument_type, normalize_aggregate,
+    qualified_name, resolve, struct_field_type, unknown_span, validate_checked_model_types,
+    validate_public_expression,
+};
 
 use crate::{
     ActionGuard, KernelModel, KernelSpec, OriginChain, OriginSite, ParamDef, TypeDef, TypeRef,
@@ -70,13 +78,6 @@ pub struct PublicKernelError {
     pub span: Option<Span>,
 }
 
-impl PublicKernelError {
-    fn with_span(mut self, span: Span) -> Self {
-        self.span.get_or_insert(span);
-        self
-    }
-}
-
 impl fmt::Display for PublicKernelError {
     fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
         formatter.write_str(&self.message)
@@ -84,6 +85,15 @@ impl fmt::Display for PublicKernelError {
 }
 
 impl std::error::Error for PublicKernelError {}
+
+impl From<TypecheckError> for PublicKernelError {
+    fn from(error: TypecheckError) -> Self {
+        Self {
+            message: error.message,
+            span: error.span,
+        }
+    }
+}
 
 fn error(message: impl Into<String>) -> PublicKernelError {
     PublicKernelError {
@@ -100,18 +110,6 @@ fn span_json(path: &str, span: Span) -> Value {
         "end_line": span.end.line,
         "end_column": span.end.column,
     })
-}
-
-fn unknown_span() -> Span {
-    let position = SourcePos {
-        offset: 0,
-        line: 1,
-        column: 1,
-    };
-    Span {
-        start: position,
-        end: position,
-    }
 }
 
 fn statement_span(statement: &Statement) -> Span {
@@ -142,40 +140,6 @@ fn type_json(ty: &TypeRef) -> Value {
     }
 }
 
-fn resolve<'a>(model: &'a KernelModel, ty: &'a TypeRef) -> Result<TypeRef, PublicKernelError> {
-    match ty {
-        TypeRef::Named(name) => match model.types.get(name) {
-            Some(TypeDef::Domain { lo, hi, .. }) => Ok(TypeRef::Range(*lo, *hi)),
-            Some(TypeDef::Enum { .. } | TypeDef::Struct { .. }) => Ok(ty.clone()),
-            None => Err(error(format!("public Kernel cannot resolve type '{name}'"))),
-        },
-        _ => Ok(ty.clone()),
-    }
-}
-
-type TypeEnv = BTreeMap<String, TypeRef>;
-type FiniteBinderCandidates = (String, Vec<(Expr, Expr)>, Option<Expr>);
-
-fn base_env(model: &KernelModel) -> TypeEnv {
-    let mut env = model.state.iter().cloned().collect::<TypeEnv>();
-    for (name, value) in &model.consts {
-        let ty = match value {
-            crate::FslValue::Bool(_) => TypeRef::Bool,
-            _ => TypeRef::Int,
-        };
-        env.entry(name.clone()).or_insert(ty);
-    }
-    for (type_name, definition) in &model.types {
-        if let TypeDef::Enum { members, .. } = definition {
-            for member in members {
-                env.entry(member.clone())
-                    .or_insert_with(|| TypeRef::Named(type_name.clone()));
-            }
-        }
-    }
-    env
-}
-
 /// Project one already-checked expression with the same typed JSON contract
 /// used inside Public Kernel v2. Sidecar contracts use this instead of
 /// publishing the historical Python-shaped AST.
@@ -191,6 +155,7 @@ pub fn public_kernel_expression(
     span: Span,
     expected_type: Option<&TypeRef>,
 ) -> Result<Value, PublicKernelError> {
+    validate_public_expression(expression, model, span, expected_type)?;
     expr_json(
         expression,
         &base_env(model),
@@ -199,284 +164,6 @@ pub fn public_kernel_expression(
         span,
         expected_type,
     )
-}
-
-pub(crate) fn validate_expression_type(
-    expr: &Expr,
-    expected: &TypeRef,
-    bindings: &[(String, TypeRef)],
-    model: &KernelModel,
-) -> Result<(), PublicKernelError> {
-    let mut env = base_env(model);
-    env.extend(bindings.iter().cloned());
-    expr_json(expr, &env, model, "", unknown_span(), Some(expected)).map(|_| ())
-}
-
-pub(crate) fn expression_binder_type(
-    binder: &Binder,
-    model: &KernelModel,
-) -> Result<TypeRef, PublicKernelError> {
-    binder_type(binder, &base_env(model), model)
-}
-
-fn qualified_name(name: &fsl_syntax::QualifiedName) -> Result<String, PublicKernelError> {
-    if name.namespace.is_some() {
-        return Err(error(
-            "qualified types must be lowered before public Kernel export",
-        ));
-    }
-    Ok(name.name.clone())
-}
-
-fn binder_type(
-    binder: &Binder,
-    env: &TypeEnv,
-    model: &KernelModel,
-) -> Result<TypeRef, PublicKernelError> {
-    match binder {
-        Binder::Typed { type_name, .. } => Ok(TypeRef::Named(qualified_name(type_name)?)),
-        Binder::Range { lo, hi, .. } => {
-            ensure_assignable(lo, &TypeRef::Int, env, model, unknown_span())?;
-            ensure_assignable(hi, &TypeRef::Int, env, model, unknown_span())?;
-            match (lo.as_ref(), hi.as_ref()) {
-                (Expr::Num(lo), Expr::Num(hi)) => Ok(TypeRef::Range(*lo, *hi)),
-                _ => Ok(TypeRef::Int),
-            }
-        }
-        Binder::Collection { collection, .. } => {
-            match resolve(model, &infer_type(collection, env, model, None)?)? {
-                TypeRef::Set(item) | TypeRef::Seq(item, _) => Ok(*item),
-                _ => Err(error("collection binder requires Set or Seq")),
-            }
-        }
-    }
-}
-
-#[allow(clippy::too_many_lines)]
-fn infer_type(
-    expr: &Expr,
-    env: &TypeEnv,
-    model: &KernelModel,
-    expected: Option<&TypeRef>,
-) -> Result<TypeRef, PublicKernelError> {
-    match expr {
-        Expr::Num(_) => match expected.map(|ty| resolve(model, ty)).transpose()? {
-            Some(ty @ (TypeRef::Int | TypeRef::Range(_, _))) => Ok(ty),
-            _ => Ok(TypeRef::Int),
-        },
-        Expr::Bool(_) | Expr::Not(_) | Expr::Is { .. } | Expr::Quantified { .. } => {
-            Ok(TypeRef::Bool)
-        }
-        Expr::None => match expected.map(|ty| resolve(model, ty)).transpose()? {
-            Some(ty @ TypeRef::Option(_)) => Ok(ty),
-            _ => Err(error("public Kernel cannot infer uncontextualized none")),
-        },
-        Expr::Some(item) => {
-            if let Some(expected) = expected
-                && matches!(resolve(model, expected)?, TypeRef::Option(_))
-            {
-                return Ok(expected.clone());
-            }
-            Ok(TypeRef::Option(Box::new(infer_type(
-                item, env, model, None,
-            )?)))
-        }
-        Expr::Set(items) => {
-            if let Some(expected) = expected
-                && matches!(resolve(model, expected)?, TypeRef::Set(_))
-            {
-                return Ok(expected.clone());
-            }
-            let first = items
-                .first()
-                .ok_or_else(|| error("public Kernel cannot infer empty Set"))?;
-            Ok(TypeRef::Set(Box::new(infer_type(first, env, model, None)?)))
-        }
-        Expr::Seq(items) => {
-            if let Some(expected) = expected
-                && matches!(resolve(model, expected)?, TypeRef::Seq(_, _))
-            {
-                return Ok(expected.clone());
-            }
-            let first = items
-                .first()
-                .ok_or_else(|| error("public Kernel cannot infer empty Seq"))?;
-            Ok(TypeRef::Seq(
-                Box::new(infer_type(first, env, model, None)?),
-                items.len(),
-            ))
-        }
-        Expr::Struct { name, .. } => Ok(TypeRef::Named(name.clone())),
-        Expr::Var(name) => env
-            .get(name)
-            .cloned()
-            .ok_or_else(|| error(format!("public Kernel cannot type identifier '{name}'"))),
-        Expr::Call { name, .. } => Err(error(format!(
-            "unlowered predicate call '{name}' in public Kernel"
-        ))),
-        Expr::Stage { .. } => Err(error("unlowered stage access in public Kernel")),
-        Expr::Index(base, _) => match resolve(model, &infer_type(base, env, model, None)?)? {
-            TypeRef::Map(_, value) | TypeRef::Relation(_, value) => Ok(*value),
-            TypeRef::Seq(item, _) => Ok(*item),
-            _ => Err(error("public Kernel index requires Map, Relation, or Seq")),
-        },
-        Expr::Field(base, field) => {
-            let TypeRef::Named(name) = infer_type(base, env, model, None)? else {
-                return Err(error("public Kernel field access requires named struct"));
-            };
-            match model.types.get(&name) {
-                Some(TypeDef::Struct { fields }) => fields
-                    .iter()
-                    .find_map(|(candidate, ty)| (candidate == field).then(|| ty.clone()))
-                    .ok_or_else(|| error(format!("unknown struct field '{name}.{field}'"))),
-                _ => Err(error("public Kernel field access requires struct")),
-            }
-        }
-        Expr::Method { receiver, name, .. } => {
-            let receiver_ty = infer_type(receiver, env, model, None)?;
-            let resolved = resolve(model, &receiver_ty)?;
-            match name.as_str() {
-                "contains" => Ok(TypeRef::Bool),
-                "size" => Ok(TypeRef::Int),
-                "add" | "remove" | "push" | "pop" => Ok(receiver_ty),
-                "head" | "at" => match resolved {
-                    TypeRef::Seq(item, _) => Ok(*item),
-                    _ => Err(error(format!("public Kernel cannot type method '{name}'"))),
-                },
-                _ => Err(error(format!("public Kernel cannot type method '{name}'"))),
-            }
-        }
-        Expr::Binary { op, .. } => {
-            if matches!(
-                op.as_str(),
-                "and" | "or" | "=>" | "==" | "!=" | "<" | "<=" | ">" | ">="
-            ) {
-                Ok(TypeRef::Bool)
-            } else {
-                Ok(TypeRef::Int)
-            }
-        }
-        Expr::Neg(_) | Expr::BinaryNamed { .. } => Ok(TypeRef::Int),
-        Expr::Aggregate { kind, .. } => Ok(
-            if matches!(kind, AggregateKind::Count | AggregateKind::Sum) {
-                TypeRef::Int
-            } else {
-                TypeRef::Bool
-            },
-        ),
-        Expr::Conditional { then_expr, .. } => infer_type(then_expr, env, model, expected),
-        Expr::UnaryNamed { name, expr, .. } => match name.as_str() {
-            "old" => infer_type(expr, env, model, expected),
-            "abs" => Ok(TypeRef::Int),
-            "rel_acyclic" | "rel_functional" | "rel_injective" => Ok(TypeRef::Bool),
-            "rel_domain" | "rel_range" => Ok(TypeRef::Set(Box::new(TypeRef::Int))),
-            _ => Err(error(format!("unsupported unary expression '{name}'"))),
-        },
-        Expr::TernaryNamed { name, .. } if name == "rel_reachable" => Ok(TypeRef::Bool),
-        Expr::TernaryNamed { name, .. } => {
-            Err(error(format!("unsupported ternary expression '{name}'")))
-        }
-    }
-}
-
-fn types_compatible(
-    actual: &TypeRef,
-    expected: &TypeRef,
-    model: &KernelModel,
-) -> Result<bool, PublicKernelError> {
-    let actual = resolve(model, actual)?;
-    let expected = resolve(model, expected)?;
-    Ok(match (&actual, &expected) {
-        (TypeRef::Int | TypeRef::Range(_, _), TypeRef::Int | TypeRef::Range(_, _))
-        | (TypeRef::Bool, TypeRef::Bool) => true,
-        (TypeRef::Named(actual), TypeRef::Named(expected)) => actual == expected,
-        (TypeRef::Set(actual), TypeRef::Set(expected))
-        | (TypeRef::Option(actual), TypeRef::Option(expected))
-        | (TypeRef::Seq(actual, _), TypeRef::Seq(expected, _)) => {
-            types_compatible(actual, expected, model)?
-        }
-        (TypeRef::Map(actual_key, actual_value), TypeRef::Map(expected_key, expected_value))
-        | (
-            TypeRef::Relation(actual_key, actual_value),
-            TypeRef::Relation(expected_key, expected_value),
-        ) => {
-            types_compatible(actual_key, expected_key, model)?
-                && types_compatible(actual_value, expected_value, model)?
-        }
-        _ => false,
-    })
-}
-
-fn ensure_assignable(
-    expr: &Expr,
-    expected: &TypeRef,
-    env: &TypeEnv,
-    model: &KernelModel,
-    span: Span,
-) -> Result<(), PublicKernelError> {
-    let resolved = resolve(model, expected).map_err(|error| error.with_span(span))?;
-    match (expr, &resolved) {
-        (Expr::None, TypeRef::Option(_)) => return Ok(()),
-        (Expr::Some(item), TypeRef::Option(expected_item)) => {
-            return ensure_assignable(item, expected_item, env, model, span);
-        }
-        (Expr::Set(items), TypeRef::Set(expected_item))
-        | (Expr::Seq(items), TypeRef::Seq(expected_item, _)) => {
-            for item in items {
-                ensure_assignable(item, expected_item, env, model, span)?;
-            }
-            return Ok(());
-        }
-        (
-            Expr::Conditional {
-                condition,
-                then_expr,
-                else_expr,
-                spans,
-            },
-            _,
-        ) => {
-            ensure_assignable(condition, &TypeRef::Bool, env, model, spans.condition)?;
-            ensure_assignable(then_expr, expected, env, model, spans.then_expr)?;
-            ensure_assignable(else_expr, expected, env, model, spans.else_expr)?;
-            return Ok(());
-        }
-        _ => {}
-    }
-    let actual = infer_type(expr, env, model, None).map_err(|error| error.with_span(span))?;
-    if types_compatible(&actual, expected, model).map_err(|error| error.with_span(span))? {
-        Ok(())
-    } else {
-        Err(error(format!(
-            "expression of type {actual:?} is not assignable to {expected:?}"
-        ))
-        .with_span(span))
-    }
-}
-
-fn extend_pattern_binding(
-    expression: &Expr,
-    env: &mut TypeEnv,
-    model: &KernelModel,
-) -> Result<(), PublicKernelError> {
-    match expression {
-        Expr::Is {
-            expr,
-            pattern: Pattern::Some(binding),
-        } => {
-            let TypeRef::Option(inner) = resolve(model, &infer_type(expr, env, model, None)?)?
-            else {
-                return Err(error("some pattern requires an Option operand"));
-            };
-            env.insert(binding.clone(), *inner);
-        }
-        Expr::Binary { left, right, .. } => {
-            extend_pattern_binding(left, env, model)?;
-            extend_pattern_binding(right, env, model)?;
-        }
-        _ => {}
-    }
-    Ok(())
 }
 
 fn binder_json(
@@ -548,217 +235,6 @@ fn binder_json(
     Ok(output)
 }
 
-fn normalize_aggregate(
-    kind: AggregateKind,
-    binder: &Binder,
-    value: Option<&Expr>,
-    env: &TypeEnv,
-    model: &KernelModel,
-) -> Result<Expr, PublicKernelError> {
-    let (name, candidates, filter) = finite_binder_candidates(binder, env, model)?;
-    let mut terms = Vec::new();
-    for (candidate, membership) in candidates {
-        let replacements = HashMap::from([(name.clone(), candidate)]);
-        let condition = aggregate_condition(membership, filter.as_ref(), &replacements);
-        let selected = match kind {
-            AggregateKind::Count | AggregateKind::Unique | AggregateKind::ExactlyOne => {
-                Expr::Num(1)
-            }
-            AggregateKind::Sum => crate::substitute_expr(
-                value
-                    .ok_or_else(|| error("sum aggregate requires a value"))?
-                    .clone(),
-                &replacements,
-            ),
-        };
-        let span = unknown_span();
-        terms.push(Expr::Conditional {
-            condition: Box::new(condition),
-            then_expr: Box::new(selected),
-            else_expr: Box::new(Expr::Num(0)),
-            spans: Box::new(ConditionalSpans {
-                condition: span,
-                then_expr: span,
-                else_expr: span,
-            }),
-        });
-    }
-    Ok(aggregate_result(kind, terms))
-}
-
-fn finite_binder_candidates(
-    binder: &Binder,
-    env: &TypeEnv,
-    model: &KernelModel,
-) -> Result<FiniteBinderCandidates, PublicKernelError> {
-    let (name, candidates, filter) = match binder {
-        Binder::Typed {
-            name, where_expr, ..
-        } => {
-            let candidates = model
-                .domain_values(&binder_type(binder, env, model)?)
-                .map_err(|model_error| error(model_error.to_string()))?
-                .into_iter()
-                .map(|candidate| Ok((value_expression(candidate)?, Expr::Bool(true))))
-                .collect::<Result<Vec<_>, PublicKernelError>>()?;
-            (name, candidates, where_expr.as_deref())
-        }
-        Binder::Range {
-            name,
-            lo,
-            hi,
-            where_expr,
-        } => {
-            let lo = static_public_int(lo, model)?;
-            let hi = static_public_int(hi, model)?;
-            let candidates = (lo..=hi)
-                .map(|candidate| (Expr::Num(candidate), Expr::Bool(true)))
-                .collect();
-            (name, candidates, where_expr.as_deref())
-        }
-        Binder::Collection {
-            name,
-            collection,
-            where_expr,
-        } => {
-            let collection_ty = resolve(model, &infer_type(collection, env, model, None)?)?;
-            let candidates = match collection_ty {
-                TypeRef::Set(element_ty) => model
-                    .domain_values(&element_ty)
-                    .map_err(|model_error| error(model_error.to_string()))?
-                    .into_iter()
-                    .map(|candidate| {
-                        let candidate = value_expression(candidate)?;
-                        let present = Expr::Method {
-                            receiver: collection.clone(),
-                            name: "contains".to_owned(),
-                            args: vec![candidate.clone()],
-                        };
-                        Ok((candidate, present))
-                    })
-                    .collect::<Result<Vec<_>, PublicKernelError>>()?,
-                TypeRef::Seq(_, capacity) => (0..capacity)
-                    .map(|index| {
-                        let index = i64::try_from(index)
-                            .map_err(|_| error("Seq capacity exceeds public integer range"))?;
-                        let candidate = Expr::Method {
-                            receiver: collection.clone(),
-                            name: "at".to_owned(),
-                            args: vec![Expr::Num(index)],
-                        };
-                        let present = Expr::Binary {
-                            op: "<".to_owned(),
-                            left: Box::new(Expr::Num(index)),
-                            right: Box::new(Expr::Method {
-                                receiver: collection.clone(),
-                                name: "size".to_owned(),
-                                args: Vec::new(),
-                            }),
-                        };
-                        Ok((candidate, present))
-                    })
-                    .collect::<Result<Vec<_>, PublicKernelError>>()?,
-                _ => return Err(error("collection binder requires Set or Seq")),
-            };
-            (name, candidates, where_expr.as_deref())
-        }
-    };
-    Ok((name.clone(), candidates, filter.cloned()))
-}
-
-fn aggregate_condition(
-    membership: Expr,
-    filter: Option<&Expr>,
-    replacements: &HashMap<String, Expr>,
-) -> Expr {
-    filter.map_or(membership.clone(), |filter| {
-        let span = unknown_span();
-        Expr::Conditional {
-            condition: Box::new(membership),
-            then_expr: Box::new(crate::substitute_expr(filter.clone(), replacements)),
-            else_expr: Box::new(Expr::Bool(false)),
-            spans: Box::new(ConditionalSpans {
-                condition: span,
-                then_expr: span,
-                else_expr: span,
-            }),
-        }
-    })
-}
-
-fn aggregate_result(kind: AggregateKind, terms: Vec<Expr>) -> Expr {
-    let total = terms
-        .into_iter()
-        .fold(Expr::Num(0), |left, right| Expr::Binary {
-            op: "+".to_owned(),
-            left: Box::new(left),
-            right: Box::new(right),
-        });
-    match kind {
-        AggregateKind::Count | AggregateKind::Sum => total,
-        AggregateKind::Unique => Expr::Binary {
-            op: "<=".to_owned(),
-            left: Box::new(total),
-            right: Box::new(Expr::Num(1)),
-        },
-        AggregateKind::ExactlyOne => Expr::Binary {
-            op: "==".to_owned(),
-            left: Box::new(total),
-            right: Box::new(Expr::Num(1)),
-        },
-    }
-}
-
-fn static_public_int(expr: &Expr, model: &KernelModel) -> Result<i64, PublicKernelError> {
-    match expr {
-        Expr::Num(value) => Ok(*value),
-        Expr::Var(name) => match model.consts.get(name) {
-            Some(crate::FslValue::Int(value)) => Ok(*value),
-            _ => Err(error(format!("'{name}' is not an integer const"))),
-        },
-        Expr::Neg(value) => static_public_int(value, model)?
-            .checked_neg()
-            .ok_or_else(|| error("integer overflow in range bound")),
-        _ => Err(error(
-            "public Kernel requires static aggregate range bounds",
-        )),
-    }
-}
-
-fn value_expression(value: crate::FslValue) -> Result<Expr, PublicKernelError> {
-    Ok(match value {
-        crate::FslValue::Int(value) => Expr::Num(value),
-        crate::FslValue::Bool(value) => Expr::Bool(value),
-        crate::FslValue::Enum { member, .. } => Expr::Var(member),
-        crate::FslValue::None => Expr::None,
-        crate::FslValue::Some(value) => Expr::Some(Box::new(value_expression(*value)?)),
-        crate::FslValue::Struct { type_name, fields } => Expr::Struct {
-            name: type_name,
-            fields: fields
-                .into_iter()
-                .map(|(name, value)| Ok((name, value_expression(value)?)))
-                .collect::<Result<Vec<_>, PublicKernelError>>()?,
-        },
-        crate::FslValue::Set(values) => Expr::Set(
-            values
-                .into_iter()
-                .map(value_expression)
-                .collect::<Result<Vec<_>, _>>()?,
-        ),
-        crate::FslValue::Seq(values) => Expr::Seq(
-            values
-                .into_iter()
-                .map(value_expression)
-                .collect::<Result<Vec<_>, _>>()?,
-        ),
-        crate::FslValue::Map(_) | crate::FslValue::Relation(_) => {
-            return Err(error(
-                "aggregate element has no public literal representation",
-            ));
-        }
-    })
-}
-
 #[allow(clippy::too_many_lines)]
 fn expr_json(
     expr: &Expr,
@@ -768,15 +244,7 @@ fn expr_json(
     span: Span,
     expected: Option<&TypeRef>,
 ) -> Result<Value, PublicKernelError> {
-    let ty = infer_type(expr, env, model, expected).map_err(|error| error.with_span(span))?;
-    if let Some(expected) = expected
-        && !types_compatible(&ty, expected, model).map_err(|error| error.with_span(span))?
-    {
-        return Err(error(format!(
-            "expression of type {ty:?} is not assignable to {expected:?}"
-        ))
-        .with_span(span));
-    }
+    let ty = expression_type(expr, env, model, span, expected)?;
     let mut output = Map::from_iter([
         ("kind".to_owned(), Value::String("unknown".to_owned())),
         ("type".to_owned(), type_json(&ty)),
@@ -824,25 +292,16 @@ fn expr_json(
         Expr::Struct { name, fields } => {
             output.insert("kind".to_owned(), json!("struct_lit"));
             output.insert("name".to_owned(), json!(name));
-            let Some(TypeDef::Struct {
-                fields: definitions,
-            }) = model.types.get(name)
-            else {
-                return Err(error(format!("unknown struct '{name}'")));
-            };
-            let definitions = definitions.iter().cloned().collect::<BTreeMap<_, _>>();
             output.insert(
                 "fields".to_owned(),
                 Value::Object(
                     fields
                         .iter()
                         .map(|(field, value)| {
-                            let expected = definitions.get(field).ok_or_else(|| {
-                                error(format!("unknown struct field '{name}.{field}'"))
-                            })?;
+                            let expected = struct_field_type(model, name, field)?;
                             Ok((
                                 field.clone(),
-                                expr_json(value, env, model, path, span, Some(expected))?,
+                                expr_json(value, env, model, path, span, Some(&expected))?,
                             ))
                         })
                         .collect::<Result<_, PublicKernelError>>()?,
@@ -862,12 +321,7 @@ fn expr_json(
             return Err(error("unlowered stage access in public Kernel"));
         }
         Expr::Index(collection, index) => {
-            let collection_ty = resolve(model, &infer_type(collection, env, model, None)?)?;
-            let key_ty = match collection_ty {
-                TypeRef::Map(key, _) | TypeRef::Relation(key, _) => *key,
-                TypeRef::Seq(_, _) => TypeRef::Int,
-                _ => return Err(error("index expression requires collection")),
-            };
+            let key_ty = index_key_type(collection, env, model)?;
             output.insert("kind".to_owned(), json!("index"));
             output.insert(
                 "collection".to_owned(),
@@ -892,12 +346,7 @@ fn expr_json(
             args,
         } => {
             let receiver_ty = resolve(model, &infer_type(receiver, env, model, None)?)?;
-            let argument_type = match (name.as_str(), &receiver_ty) {
-                ("contains" | "add" | "remove", TypeRef::Set(item))
-                | ("contains" | "push", TypeRef::Seq(item, _)) => Some(item.as_ref()),
-                ("at", TypeRef::Seq(_, _)) => Some(&TypeRef::Int),
-                _ => None,
-            };
+            let argument_type = method_argument_type(name, &receiver_ty);
             output.insert("kind".to_owned(), json!("method"));
             output.insert(
                 "receiver".to_owned(),
@@ -908,25 +357,13 @@ fn expr_json(
                 "arguments".to_owned(),
                 Value::Array(
                     args.iter()
-                        .map(|arg| expr_json(arg, env, model, path, span, argument_type))
+                        .map(|arg| expr_json(arg, env, model, path, span, argument_type.as_ref()))
                         .collect::<Result<_, _>>()?,
                 ),
             );
         }
         Expr::Binary { op, left, right } => {
-            let (left_ty, right_ty) = if matches!(op.as_str(), "and" | "or" | "=>") {
-                (TypeRef::Bool, TypeRef::Bool)
-            } else if matches!(
-                op.as_str(),
-                "+" | "-" | "*" | "/" | "%" | "<" | "<=" | ">" | ">="
-            ) {
-                (TypeRef::Int, TypeRef::Int)
-            } else {
-                infer_type(left, env, model, None).map_or_else(
-                    |_| infer_type(right, env, model, None).map(|ty| (ty.clone(), ty)),
-                    |ty| Ok((ty.clone(), ty)),
-                )?
-            };
+            let (left_ty, right_ty) = binary_operand_types(op, left, right, env, model)?;
             output.insert("kind".to_owned(), json!("binary"));
             output.insert("operator".to_owned(), json!(op));
             let left_json = expr_json(left, env, model, path, span, Some(&left_ty))?;
@@ -1123,39 +560,6 @@ fn expr_json(
     Ok(Value::Object(output))
 }
 
-fn lvalue_type(
-    target: &LValue,
-    env: &TypeEnv,
-    model: &KernelModel,
-) -> Result<TypeRef, PublicKernelError> {
-    match target {
-        LValue::Var(name) => env
-            .get(name)
-            .cloned()
-            .ok_or_else(|| error(format!("unknown update target '{name}'"))),
-        LValue::Index(name, _) => match resolve(
-            model,
-            env.get(name)
-                .ok_or_else(|| error(format!("unknown update target '{name}'")))?,
-        )? {
-            TypeRef::Map(_, value) | TypeRef::Relation(_, value) => Ok(*value),
-            _ => Err(error("indexed update target requires Map or Relation")),
-        },
-        LValue::Field(base, field) => {
-            let TypeRef::Named(name) = lvalue_type(base, env, model)? else {
-                return Err(error("field update target requires named struct"));
-            };
-            match model.types.get(&name) {
-                Some(TypeDef::Struct { fields }) => fields
-                    .iter()
-                    .find_map(|(candidate, ty)| (candidate == field).then(|| ty.clone()))
-                    .ok_or_else(|| error(format!("unknown struct field '{name}.{field}'"))),
-                _ => Err(error("field update target requires struct")),
-            }
-        }
-    }
-}
-
 fn lvalue_json(
     target: &LValue,
     env: &TypeEnv,
@@ -1238,163 +642,6 @@ fn statement_json(
                 "binder":binder_json(binder,env,model,path,*span)?,
                 "statements":statements.iter().map(|item|statement_json(item,&local,model,path)).collect::<Result<Vec<_>,_>>()?
             }))
-        }
-    }
-}
-
-/// Run the public Kernel expression/type checker for one normalized statement.
-///
-/// Inline state initializers lower to ordinary assignments, so keeping this
-/// entry point beside `statement_json` prevents their validation rules from
-/// drifting away from the existing normalized Kernel contract.
-pub(crate) fn validate_statement_types(
-    statement: &Statement,
-    model: &KernelModel,
-) -> Result<(), PublicKernelError> {
-    validate_statement_assignments(statement, &base_env(model), model)
-}
-
-/// Run the shared expression/type checker over every non-init model expression.
-pub(crate) fn validate_model_expression_types(
-    model: &KernelModel,
-) -> Result<(), PublicKernelError> {
-    let env = base_env(model);
-    for action in &model.actions {
-        let mut local = env.clone();
-        for param in &action.params {
-            match param {
-                ParamDef::Typed { name, ty } => {
-                    local.insert(name.clone(), ty.clone());
-                }
-                ParamDef::Range { name, lo, hi } => {
-                    local.insert(name.clone(), TypeRef::Range(*lo, *hi));
-                }
-            }
-        }
-        for guard in &action.guards {
-            match guard {
-                ActionGuard::Requires(expr) => {
-                    expr_json(expr, &local, model, "", action.span, Some(&TypeRef::Bool))?;
-                    extend_pattern_binding(expr, &mut local, model)?;
-                }
-                ActionGuard::Let(name, expr) => {
-                    expr_json(expr, &local, model, "", action.span, None)?;
-                    local.insert(name.clone(), infer_type(expr, &local, model, None)?);
-                }
-            }
-        }
-        for statement in &action.statements {
-            statement_json(statement, &local, model, "")?;
-        }
-        for expr in &action.ensures {
-            expr_json(expr, &local, model, "", action.span, Some(&TypeRef::Bool))?;
-        }
-    }
-    for property in model
-        .invariants
-        .iter()
-        .chain(&model.transitions)
-        .chain(&model.reachables)
-    {
-        expr_json(
-            &property.expr,
-            &env,
-            model,
-            "",
-            property.span,
-            Some(&TypeRef::Bool),
-        )?;
-    }
-    for leadsto in &model.leadstos {
-        let mut local = env.clone();
-        for binder in &leadsto.binders {
-            binder_json(binder, &local, model, "", leadsto.span)?;
-            let name = match binder {
-                Binder::Typed { name, .. }
-                | Binder::Range { name, .. }
-                | Binder::Collection { name, .. } => name,
-            };
-            local.insert(name.clone(), binder_type(binder, &local, model)?);
-        }
-        expr_json(
-            &leadsto.before,
-            &local,
-            model,
-            "",
-            leadsto.span,
-            Some(&TypeRef::Bool),
-        )?;
-        expr_json(
-            &leadsto.after,
-            &local,
-            model,
-            "",
-            leadsto.span,
-            Some(&TypeRef::Bool),
-        )?;
-        if let Some(expr) = &leadsto.decreases {
-            expr_json(expr, &local, model, "", leadsto.span, Some(&TypeRef::Int))?;
-        }
-    }
-    if let Some(expr) = &model.terminal {
-        let position = SourcePos {
-            offset: 0,
-            line: 1,
-            column: 1,
-        };
-        expr_json(
-            expr,
-            &env,
-            model,
-            "",
-            Span {
-                start: position,
-                end: position,
-            },
-            Some(&TypeRef::Bool),
-        )?;
-    }
-    Ok(())
-}
-
-fn validate_statement_assignments(
-    statement: &Statement,
-    env: &TypeEnv,
-    model: &KernelModel,
-) -> Result<(), PublicKernelError> {
-    match statement {
-        Statement::Assign { .. } => statement_json(statement, env, model, "").map(|_| ()),
-        Statement::If {
-            then_statements,
-            else_statements,
-            ..
-        } => then_statements
-            .iter()
-            .chain(else_statements)
-            .try_for_each(|item| validate_statement_assignments(item, env, model)),
-        Statement::ForAll {
-            binder, statements, ..
-        } => {
-            let (name, ty) = match binder {
-                Binder::Typed {
-                    name, type_name, ..
-                } => (name, TypeRef::Named(qualified_name(type_name)?)),
-                Binder::Range { name, .. } => (name, TypeRef::Int),
-                Binder::Collection {
-                    name, collection, ..
-                } => {
-                    let collection_ty = resolve(model, &infer_type(collection, env, model, None)?)?;
-                    let (TypeRef::Set(item) | TypeRef::Seq(item, _)) = collection_ty else {
-                        return Err(error("collection binder requires Set or Seq"));
-                    };
-                    (name, *item)
-                }
-            };
-            let mut local = env.clone();
-            local.insert(name.clone(), ty);
-            statements
-                .iter()
-                .try_for_each(|item| validate_statement_assignments(item, &local, model))
         }
     }
 }
@@ -2022,6 +1269,7 @@ pub fn public_kernel_contract(
             "public Kernel v1 cannot preserve component source filenames after compose lowering",
         ));
     }
+    validate_checked_model_types(model)?;
     let env = base_env(model);
     let kinds = source_property_kinds(kernel);
     let constants = model
